requirements.txt
streamlit>=1.33
pyyaml>=6.0